Move bg96 code to source folder.
